[zeta Module CLEAN UP OPERATIO]
