+++ zeta/__init__.py
@@ -1,36 +1,13 @@
-import logging
-import os
-import warnings
+from zeta.utils.disable_logging import disable_warnings_and_logs
 
-# disable warnings
+disable_warnings_and_logs()
 
-warnings.filterwarnings("ignore")
-
-# disable tensorflow warnings
-
-os.environ["TF_CPP_MIN_LOG_LEVEL"] = "2"
-
-# disable bnb warnings and others
-
-logging.getLogger().setLevel(logging.WARNING)
-
-
-class CustomFilter(logging.Filter):
-    def filter(self, record):
-        msg = "Created a temporary directory at"
-        return msg not in record.getMessage()
-
-
-logger = logging.getLogger()
-f = CustomFilter()
-logger.addFilter(f)
-
-from zeta.nn import *
-from zeta.models import *
-from zeta.utils import *
-from zeta.training import *
-from zeta.tokenizers import *
-from zeta.rl import *
-from zeta.optim import *
-from zeta.ops import *
-from zeta.quant import *
+from zeta.nn import *  # noqa: F403, E402
+from zeta.models import *  # noqa: F403, E402
+from zeta.utils import *  # noqa: F403, E402
+from zeta.training import *  # noqa: F403, E402
+from zeta.tokenizers import *  # noqa: F403, E402
+from zeta.rl import *  # noqa: F403, E402
+from zeta.optim import *  # noqa: F403, E402
+from zeta.ops import *  # noqa: F403, E402
+from zeta.quant import *  # noqa: F403, E402
